fix(text3d): clean 503 'unavailable' state + extract shared useModelViewer
Two fast-follows from the text-to-3D preview review:
- 503 (server has no TRIPO_API_KEY yet) now maps to a quiet 'unavailable'
  panel state instead of surfacing the raw JSON error; the input is hidden so
  users can't retry into a wall.
- Extract the lazy-CDN useModelViewer hook (was duplicated in GallerySection
  and PreviewConceptPanel) into src/funnel/hooks/useModelViewer.ts; both now
  import it.
